+++ src/core/src/lib/extensions/field-form/field-form.ts
@@ -8,7 +8,7 @@ import {
   AsyncValidatorFn,
 } from '@angular/forms';
 import { getFieldValue, defineHiddenProp, isNil } from '../../utils';
-import { registerControl, findControl, updateValidity as updateControlValidity } from './utils';
+import { registerControl, findControl, updateValidity } from './utils';
 import { of } from 'rxjs';
 
 /** @experimental */
@@ -41,20 +41,7 @@ export class FieldFormExtension implements FormlyExtension {
     }
 
     this.root = null;
-    const fieldsToUpdate = this.setValidators(field);
-    if (fieldsToUpdate.length === 0) {
-      return;
-    }
-
-    if (fieldsToUpdate.length === 1) {
-      let control = fieldsToUpdate[0].formControl;
-      while (control) {
-        (control as any)._updateTreeValidity({ onlySelf: true });
-        control = control.parent;
-      }
-    } else {
-      (field.form as any)._updateTreeValidity();
-    }
+    this.setValidators(field);
   }
 
   private addFormControl(field: FormlyFieldConfigCache) {
@@ -74,27 +61,21 @@ export class FieldFormExtension implements FormlyExtension {
   }
 
   private setValidators(field: FormlyFieldConfigCache) {
-    let updateValidity = false;
+    let markForCheck = false;
+    field.fieldGroup?.forEach((f) => f && this.setValidators(f) && (markForCheck = true));
+
     if (!isNil(field.key) || !field.parent || (isNil(field.key) && !field.fieldGroup)) {
       const { formControl: c } = field;
       const disabled = field.templateOptions ? field.templateOptions.disabled : false;
       if (field.key && c) {
         if (disabled && c.enabled) {
           c.disable({ emitEvent: false, onlySelf: true });
-          if (!c.parent) {
-            updateControlValidity(c);
-          } else {
-            updateValidity = true;
-          }
+          markForCheck = true;
         }
 
         if (!disabled && c.disabled) {
           c.enable({ emitEvent: false, onlySelf: true });
-          if (!c.parent) {
-            updateControlValidity(c);
-          } else {
-            updateValidity = true;
-          }
+          markForCheck = true;
         }
       }
 
@@ -110,25 +91,15 @@ export class FieldFormExtension implements FormlyExtension {
           return v ? v(c) : of(null);
         });
 
-        if (!c.parent) {
-          updateControlValidity(c);
-        } else {
-          updateValidity = true;
-        }
+        markForCheck = true;
+      }
+
+      if (markForCheck) {
+        updateValidity(c, true);
       }
     }
 
-    const fieldsToUpdate = updateValidity ? [field] : [];
-    (field.fieldGroup || []).forEach((f) => {
-      if (f) {
-        const childrenToUpdate = this.setValidators(f);
-        if (!updateValidity) {
-          fieldsToUpdate.push(...childrenToUpdate);
-        }
-      }
-    });
-
-    return fieldsToUpdate;
+    return markForCheck;
   }
 
   private mergeValidators<T>(field: FormlyFieldConfigCache, type: '_validators' | '_asyncValidators'): T[] {